Refactor : refactor Gson data name (#23256)
